Jd/develop power simulations

+++ src/models/topological_elements.jl
@@ -1,7 +1,7 @@
 #TODO: Update to named tuples when Julia 0.7 becomes available
 
 struct Bus <: PowerSystemDevice
-    number::Int
+    number::Int64
     name::String
     bustype::Union{String,Nothing} # [PV, PQ, SF]
     angle::Union{Float64,Nothing} # [degrees]
